% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desx.R
\docType{data}
\name{desx}
\alias{desx}
\title{GBM clinical annotation data}
\description{
This is the clinical annotation data from the GBM dataset, it contains the class of the tumour
which is one of: classical, mesenchymal, neural, proneural. It is a data frame with 2 columns and 50 rows.
}
\references{
Verhaak, Roel GW, et al. "Integrated genomic analysis identifies clinically 
relevant subtypes of glioblastoma characterized by abnormalities in PDGFRA, IDH1, EGFR, 
and NF1." Cancer cell 17.1 (2010): 98-110.
}
\author{
Chris John \email{chris.r.john86@gmail.com}
}
\keyword{data}
