% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefilter.R
\name{featurefilter}
\alias{featurefilter}
\title{featurefilter: A function for filtering genes based on the coefficient of variance (CV)}
\usage{
featurefilter(mydata, mypercentile = 75)
}
\arguments{
\item{mydata}{Data frame or matrix or M3C results object: if dataframe/matrix should have samples as columns and rows as features}

\item{mypercentile}{Numerical value: CV percentile by which to filter the features, e.g. 90 indicates top 10% based on CV}
}
\value{
A filtered (fewer rows) data frame
}
\description{
featurefilter: A function for filtering genes based on the coefficient of variance (CV)
}
\examples{
filtered <- featurefilter(mydata,mypercentile=90)
}
