% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{GBM expression data}
\description{
This is the expression data from the GBM dataset
}
\references{
Verhaak, Roel GW, et al. "Integrated genomic analysis identifies clinically 
relevant subtypes of glioblastoma characterized by abnormalities in PDGFRA, IDH1, EGFR, 
and NF1." Cancer cell 17.1 (2010): 98-110.
}
\author{
Chris John \email{chris.r.john86@gmail.com}
}
\keyword{data}
