% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{pca: A principal component analysis function}
\usage{
pca(mydata, K = FALSE, printres = FALSE)
}
\arguments{
\item{mydata}{Data frame or matrix or M3C results object: if dataframe/matrix should have samples as columns and rows as features}

\item{K}{Numerical value: if running on the M3C results object, which value was the optimal K?}

\item{printres}{Logical flag: whether to print the PCA into current directory}
}
\value{
A PCA plot object
}
\description{
pca: A principal component analysis function
}
\examples{
PCA <- pca(mydata)
}
