% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefilter.R
\name{featurefilter}
\alias{featurefilter}
\title{featurefilter: A function for filtering features based on the coefficient of variance}
\usage{
featurefilter(mydata, percentile = 10)
}
\arguments{
\item{mydata}{Data frame: should have samples as columns and rows as features}

\item{percentile}{Numerical value: the top X percent most variable features should be kept}
}
\value{
A filtered data frame
}
\description{
featurefilter: A function for filtering features based on the coefficient of variance
}
\examples{
filtered <- featurefilter(mydata,percentile=10)
}
