% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3C.R
\name{M3C}
\alias{M3C}
\title{M3C: Monte Carlo Consensus Clustering}
\usage{
M3C(mydata, montecarlo = TRUE, cores = 1, iters = 100, maxK = 10,
  des = NULL, ref_method = "reverse-pca", repsref = 100, repsreal = 100,
  clusteralg = "pam", distance = "euclidean", pacx1 = 0.1, pacx2 = 0.9,
  printres = FALSE, printheatmaps = FALSE, showheatmaps = FALSE,
  seed = NULL)
}
\arguments{
\item{mydata}{Data frame or matrix: Contains the data, with samples as columns and rows as features}

\item{montecarlo}{Logical flag: whether to run the monte carlo simulation or not (recommended: TRUE)}

\item{cores}{Numerical value: how many cores to split the monte carlo simulation over}

\item{iters}{Numerical value: how many monte carlo iterations to perform (default: 100, recommended: 100-1000)}

\item{maxK}{Numerical value: the maximum number of clusters to test for, K (default: 10)}

\item{des}{Data frame: contains annotation data for the input data for automatic reordering (optional)}

\item{ref_method}{Character string: refers to which reference method to use (recommended: leaving this alone)}

\item{repsref}{Numerical value: how many reps to use for the monte carlo reference data (suggest 100)}

\item{repsreal}{Numerical value: how many reps to use for the real data (recommended: 100)}

\item{clusteralg}{String: dictates which algorithm to use for M3C (recommended: leaving this alone)}

\item{distance}{String: dictates which distance metric to use for M3C (recommended: leaving this alone)}

\item{pacx1}{Numerical value: The 1st x co-ordinate for calculating the pac score from the CDF (default: 0.1)}

\item{pacx2}{Numerical value: The 2nd x co-ordinate for calculating the pac score from the CDF (default: 0.9)}

\item{printres}{Logical flag: whether to print all results into current directory}

\item{printheatmaps}{Logical flag: whether to print all the heatmaps into current directory}

\item{showheatmaps}{Logical flag: whether to show the heatmaps on screen (can be slow)}

\item{seed}{Numerical value: fixes the seed if you want to repeat results, set the seed to 123 for example here}
}
\value{
A list: containing 1) the stability results and 2) all the output data (in another list) 
(see vignette for more details on how to access)
}
\description{
M3C: Monte Carlo Consensus Clustering
}
\examples{
res <- M3C(mydata, cores=1, iters=100, ref_method = 'reverse-pca', montecarlo = TRUE,printres = FALSE, 
maxK = 10, showheatmaps = FALSE, repsreal = 100, repsref = 100,printheatmaps = FALSE, seed = 123, des = desx)
}
