% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{pca: A principal component analysis function}
\usage{
pca(mydata, K = FALSE, printres = FALSE, labels = FALSE,
  text = FALSE, axistextsize = 18, legendtextsize = 18,
  dotsize = 5, textlabelsize = 4)
}
\arguments{
\item{mydata}{Data frame or matrix or M3C results object: if dataframe/matrix should have samples as columns and rows as features}

\item{K}{Numerical value: if running on the M3C results object, which value was the optimal K?}

\item{printres}{Logical flag: whether to print the PCA into current directory}

\item{labels}{Character vector: if we want to just label with gender for example}

\item{text}{Character vector: if we wanted to label the samples with text IDs to look for outliers}

\item{axistextsize}{Numerical value: axis text size}

\item{legendtextsize}{Numerical value: legend text size}

\item{dotsize}{Numerical value: dot size}

\item{textlabelsize}{Numerical value: text inside plot label size}
}
\value{
A PCA plot object
}
\description{
pca: A principal component analysis function
}
\examples{
PCA <- pca(mydata)
}
