% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{tsne}
\alias{tsne}
\title{tsne: A tsne function}
\usage{
tsne(mydata, K = FALSE, labels = FALSE, perplex = 15,
  printres = FALSE, seed = FALSE, axistextsize = 18,
  legendtextsize = 18, dotsize = 5, textlabelsize = 4)
}
\arguments{
\item{mydata}{Data frame or matrix or M3C results (list) object: if dataframe/matrix should have samples as columns and rows as features}

\item{K}{Numerical value: if running on the M3C results object, which value was the optimal K? Needs manual input from user.}

\item{labels}{Factor: if we want to just display gender for example, only for when running without K parameter and with a matrix or data frame}

\item{perplex}{Numerical value: this is the perplexity parameter for tsne, it usually requires adjusting for each dataset}

\item{printres}{Logical flag: whether to print the plot into current directory}

\item{seed}{Numerical value: to repeat the results exactly, setting seed is required}

\item{axistextsize}{Numerical value: axis text size}

\item{legendtextsize}{Numerical value: legend text size}

\item{dotsize}{Numerical value: dot size}

\item{textlabelsize}{Numerical value: text inside plot label size}
}
\value{
A tsne plot object
}
\description{
tsne: A tsne function
}
\examples{
TSNE <- tsne(mydata,perplex=15)
}
